requirements.txt
flask==3.1.1
openpyxl==3.1.5
python-docx==1.2.0